+++ app_flomart_main/app_flomart_main.py
@@ -2,8 +2,8 @@
 """
 FLOMART - FLOod MAp in Real Time
 
-__date__ = '20241129'
-__version__ = '2.2.0'
+__date__ = '20241215'
+__version__ = '2.2.1'
 __author__ =
         'Fabio Delogu (fabio.delogu@cimafoundation.org)',
         'Matteo Darienzo (matteo.darienzo@cimafoundation.org)',
@@ -18,6 +18,7 @@ General command line:
 python app_flomart_main.py -settings_file configuration.json -time "YYYY-MM-DD HH:MM"
 
 Version(s):
+20241215 (2.2.1) --> Add the method to dump reference areas, fix bugs in geographical destination boundaries
 20241129 (2.2.0) --> Add safe memory mode to manage the ram memory usage, extend codes to manage the info domain file, fix bugs
 20240530 (2.1.1) --> Fix bugs to manage the case of missing data in the observed dataset (no links mode)
 20231106 (2.1.0) --> Add mode using Q_T file(s), bugs fixing and code refactoring (to adapt new sources and format)
@@ -59,8 +60,8 @@ log_stream = logging.getLogger(logger_name)
 
 # -------------------------------------------------------------------------------------
 # Algorithm information
-alg_version = '2.2.0'
-alg_release = '2024-11-29'
+alg_version = '2.2.1'
+alg_release = '2024-12-15'
 alg_name = 'FLOMART'
 # -------------------------------------------------------------------------------------